billing/app/requirements.txt
Flask
Flask-RESTful
mysql-connector